requirements.txt
griptape @ git+https://github.com/griptape-ai/griptape@dev
python-dotenv